Able to perform consecutive calculations.
All buttons except for clear and delete buttons are given the
setEquation functions to decide which of the three variables, operand1,
operand2 and operator will be modified. Within the function, it is
separated into two functions deciding which variable to modify and what
action to take to decrease the amount of code within the function.